Move non-Windows code behind build tags
Moved non-windows code to new non-windows file so it could be compiled on Windows.
